feat: add iss to oidc data (#99)
* feat: add iss to oidc data

add iss to oidc data

* chore: bump lock

* chore: remove unused parse module

remove unused parse module
