+++ packages/auth-server/stores/local-node.json
@@ -1,9 +1,9 @@
 {
-  "oidcKeyRegistry": "0x9d1BC7880da43C6f12CB4cD581EF4CEdD884a9ce",
+  "oidcKeyRegistry": "0xe92E4735683801ef87EAe79e1D8fC75Efa9ca117",
   "session": "0xa5980997eB1cb8942d6aAf54b87c311D0af05d3C",
   "passkey": "0xc0796b37178e855198d293779339b2adC697565B",
   "recovery": "0xDe95b35EBeDBc3ec62f3298C642cdA8Ee5E9C614",
-  "recoveryOidc": "0xfd67726e473ed143d7BcBF8e03EaC362Fbc231C7",
+  "recoveryOidc": "0x485df8Cd344462008Fa7bC44f98391c3DB344Ee0",
   "accountFactory": "0x4e75295Fd05350150DAD57C991117765F1BCE0C9",
-  "accountPaymaster": "0x71Aa4Edf60D87e174C7a69f36661C66302805ee0"
+  "accountPaymaster": "0x528b3aAa486D4888B93eC15e59b2363c1aDBF24C"
 }
