+++ packages/sdk/src/client/oidc/index.ts
@@ -1,3 +1,2 @@
 export * from "./account.js";
 export * from "./client.js";
-export * from "./parse.js";
