+++ packages/sdk/src/client/recovery/actions/oidc.ts
@@ -19,7 +19,14 @@ import {
 
 import { OidcRecoveryModuleAbi } from "../../../abi/index.js";
 import { noThrow } from "../../../utils/helpers.js";
-import type { OidcData } from "../../oidc/index.js";
+
+export type OidcData = {
+  oidcDigest: Hex;
+  iss: string;
+  readyToRecover: boolean;
+  pendingPasskeyHash: Hex;
+  recoverNonce: Hex;
+};
 
 export type AddOidcAccountArgs = {
   contracts: {
@@ -43,10 +50,13 @@ export const addOidcAccount = async <
   const encodedOidcData = encodeAbiParameters(
     [
       {
-        type: "bytes32", name: "key",
+        type: "tuple", name: "OidcData", components: [
+          { type: "bytes32", name: "oidcDigest" },
+          { type: "string", name: "iss" },
+        ],
       },
     ],
-    [args.oidcData.oidcDigest],
+    [args.oidcData],
   );
 
   const callData = encodeFunctionData({
